Add icons to rooms and sessions without recording.
Rooms without video equipment get marked in the header line of the
tables. Sessions without recording (but equipment in the room) get
marked in the cell. In both cases, the page of the session itself
contains both an icon and a textual description.

All icons have a title attribute to make browsers show a tooltip.